правки ошибок

+++ race/views.py
@@ -81,7 +81,8 @@ def protocol_all(request, race_code):
     for user in users:
         protocols[user.id][0] = user.username
         protocols[user.id][1] = Protocol.objects.filter(race=race, user=user).last()
-        protocols[user.id][2] = UserRaceLink.objects.filter(user = user, race = race).first()
+        if request.user.is_authenticated():
+            protocols[user.id][2] = UserRaceLink.objects.filter(user = user, race = race).first()
 
     args['protocols'] = protocols
     args['race_code'] = race_code
@@ -96,30 +97,34 @@ def race_select(request):
     races = [[0 for x in range(2)] for y in range(1000)]
     for race in Race.objects.all():
         races[race.id][0] = race
-        races[race.id][1] = UserRaceLink.objects.filter(race=race, user=request.user)
+        if request.user.is_authenticated():
+            races[race.id][1] = UserRaceLink.objects.filter(race=race, user=request.user)
     args['races'] = races
     return render(request, 'race/race_select.html', args )
 
 def race_home(request, race_code):
     args = {}
     args['race'] = Race.objects.get(title = race_code)
-    args['registration'] = UserRaceLink.objects.filter(race=args['race'], user=request.user)
+    if request.user.is_authenticated():
+        args['registration'] = UserRaceLink.objects.filter(race=args['race'], user=request.user)
     return render(request, 'race/race_home.html', args )
 
 def registration(request, race_code):
-    race = Race.objects.get(title = race_code)
+    if request.user.is_authenticated():
+        race = Race.objects.get(title = race_code)
 
-    link = UserRaceLinkForm().save(commit=False)
-    link.user = request.user
-    link.race_id = race.id
+        link = UserRaceLinkForm().save(commit=False)
+        link.user = request.user
+        link.race_id = race.id
 
-    link.save()
+        link.save()
 
     return redirect('race_select')
 
 def registration_remove(request, race_code):
-    race = Race.objects.get(title = race_code)
-    link = UserRaceLink.objects.filter(user = request.user, race = race)
-    link.delete()
+    if request.user.is_authenticated():
+        race = Race.objects.get(title = race_code)
+        link = UserRaceLink.objects.filter(user = request.user, race = race)
+        link.delete()
 
     return redirect('race_select')
